requirements.txt
python-nmap>=0.7.1
dnspython>=2.1.0
requests>=2.26.0
python-whois>=0.7.3
beautifulsoup4>=4.9.3
selenium>=4.1.0
aiohttp>=3.8.1
python-dotenv>=0.19.0
slack-sdk>=3.11.0
slack-bolt>=1.14.0
rich>=10.12.0 